requirements.txt
flask
flask-login
flask-mail
flask-wtf
python-dotenv
razorpay
sqlite3
Flask
Flask-Mail
python-dotenv
razorpay

